Port SSE Code to Arm Neon.
There are a few small differences between the branches, but it
seems to work well enough.
