+++ cbench/bench_atoi.c
@@ -1,5 +1,5 @@
 #include "anemone_atoi.h"
-#include "anemone_atoi_sse.h"
+#include "anemone_atoi_v128.h"
 #include <stdlib.h>
 
 #define NUM_BENCHES 1000000
